glossary/app/requirements.txt
streamlit
google.cloud.logging
google-cloud-translate
st-files-connection
gcsfs